requirements.txt
asgiref==3.8.1
dj-config-url==0.1.1
Django==5.2.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
